Development

+++ frontend/src/pages/PeoplePage/PeoplePage.jsx
@@ -1,14 +1,18 @@
 import "./PeoplePage.css";
 import "../../styles/form.css";
 import "../../styles/popups.css";
+import "../../styles/stateMsg.css";
 import PeopleTable from "../../components/People/PeopleTable";
-import { useEffect, useState } from "react";
+import Searchbar from "../../components/common/Searchbar";
+import DeleteConfirm from "../../components/common/DeleteConfirm";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { apiFetch } from "../../api";
+import PeopleFilterBar from "../../components/People/PeopleFilterBar";
 
 export default function PeoplePage() {
   const [people, setPeople] = useState([]);
   const [currentPage, setCurrentPage] = useState(1);
-  const [total, setTotal] = useState(0);
+  const [, setTotal] = useState(0);
   const [totalPages, setTotalPages] = useState(1);
   const limit = 5; // show 5 people on each page
 
@@ -16,42 +20,145 @@ export default function PeoplePage() {
   const [showAddForm, setShowAddForm] = useState(false);
   const [showMsg, setShowMsg] = useState(false);
   const [msg, setMsg] = useState("");
+  const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
 
-  // patient fields
+  // people fields
   const [name, setName] = useState("");
   const [email, setEmail] = useState("");
   const [phone, setPhone] = useState("");
   const [notes, setNotes] = useState("");
   const [status, setStatus] = useState(0);
+  const [role, setRole] = useState("patient");
 
   const [formTitle, setFormTitle] = useState("");
   const [isEditMode, setIsEditMode] = useState(false);
   const [editingPersonId, setEditingPersonId] = useState(null);
+  const [selectedPerson, setSelectedPerson] = useState(null);
+
+  // role tabs
+  const [roleTab, setRoleTab] = useState("patient");
+
+  // search bar
+  const [search, setSearch] = useState("");
+
+  // states
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState("");
+
+  // filter
+  const [filterStatus, setFilterStatus] = useState("");
+  const [filterName, setFilterName] = useState("");
+
+  // import people
+  const fileInputRef = useRef(null);
+
+  // open file picker
+  const handleImportClick = () => {
+    fileInputRef.current?.click();
+  };
+
+  // upload selected csv file
+  const handleImportFileChange = async (e) => {
+    const file = e.target.files?.[0];
+
+    if (!file) return;
+
+    try {
+      const formData = new FormData();
+      formData.append("file", file);
+
+      const token = localStorage.getItem("token");
+
+      const response = await fetch("api/people/import", {
+        method: "POST",
+        headers: {
+          Authorization: `Bearer ${token}`,
+        },
+        body: formData,
+      });
+
+      const data = await response.json();
+
+      if (!response.ok) {
+        throw new Error(data.error || "Import failed");
+      }
+
+      setMsg(
+        `Import completed: ${data.insertedCount} inserted, ${data.failedCount} failed`,
+      );
+      setShowMsg(true);
+
+      setTimeout(() => {
+        setShowMsg(false);
+      }, 2000);
+
+      // refresh people list
+      fetchPeople();
+    } catch (err) {
+      console.error("Import people error:", err);
+      setMsg("Failed to import people");
+      setShowMsg(true);
+
+      setTimeout(() => {
+        setShowMsg(false);
+      }, 2000);
+    }
+
+    // clear input value so same file can be selected again
+    e.target.value = "";
+  };
+
+  // handle delete confirm
+  const handleOpenDeleteConfirm = (person) => {
+    setSelectedPerson(person);
+    setShowDeleteConfirm(true);
+  };
+
+  // handle delete cancel
+  const handleCancelDelete = () => {
+    setShowDeleteConfirm(false);
+    setSelectedPerson(null);
+  };
+
+  // filter clear
+  const handleFilterClear = () => {
+    setFilterStatus("");
+    setCurrentPage(1);
+    setFilterName("");
+  };
+
+  // fetch people
+  const fetchPeople = useCallback(async () => {
+    try {
+      // set is loading
+      setLoading(true);
+      setError("");
+
+      const r = await apiFetch(
+        `/api/people?limit=${limit}&page=${currentPage}&role=${roleTab}&search=${search}&filterStatus=${filterStatus}&filterName=${filterName}`,
+      );
+      if (!r.ok) throw new Error(`HTTP ${r.status}`);
+      const data = await r.json();
+
+      // set data
+      setPeople(data.data);
+      setTotal(data.total);
+      setTotalPages(data.totalPages);
+    } catch (e) {
+      console.error("fetch failed:", e);
+    } finally {
+      setLoading(false);
+    }
+  }, [currentPage, roleTab, search, filterStatus, filterName]);
 
   useEffect(() => {
-    fetchPatients();
-  }, [currentPage]);
-
-  // fetch patients
-  const fetchPatients = async () => {
-    apiFetch(`/api/people?limit=${limit}&page=${currentPage}`)
-      .then((r) => {
-        if (!r.ok) throw new Error(`HTTP ${r.status}`);
-        return r.json();
-      })
-      .then((data) => {
-        setPeople(data.data);
-        setTotal(data.total);
-        setTotalPages(data.totalPages);
-      })
-      .catch((e) => console.error("fetch failed:", e));
-  };
+    fetchPeople();
+  }, [fetchPeople]);
 
   const handleSubmit = async (e) => {
     e.preventDefault();
 
     let response;
-
     if (isEditMode) {
       response = await apiFetch(`/api/people/${editingPersonId}`, {
         method: "PUT",
@@ -68,12 +175,13 @@ export default function PeoplePage() {
       });
     } else {
       // send req & get res
-      response = await apiFetch("/api/people/add-patient", {
+      response = await apiFetch("/api/people/add-person", {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
         },
         body: JSON.stringify({
+          role,
           name,
           email,
           phone,
@@ -87,16 +195,16 @@ export default function PeoplePage() {
     // set msg
     if (response.ok) {
       isEditMode
-        ? setMsg("Patient edited successfully")
-        : setMsg("Patient added successfully");
+        ? setMsg("Person edited successfully")
+        : setMsg("Person added successfully");
     } else {
       isEditMode
-        ? setMsg("Failed to edit patient")
-        : setMsg("Failed to add patient");
+        ? setMsg("Failed to edit person")
+        : setMsg("Failed to add person");
     }
 
     // refresh people list
-    await fetchPatients();
+    await fetchPeople();
 
     setShowMsg(true);
 
@@ -110,9 +218,10 @@ export default function PeoplePage() {
     //show msg
   };
 
-  const handleDelete = async (personId) => {
+  const handleDelete = async () => {
+    if (!selectedPerson) return;
     // send req & get res
-    const response = await apiFetch(`/api/people/${personId}`, {
+    const response = await apiFetch(`/api/people/${selectedPerson.id}`, {
       method: "DELETE",
     });
 
@@ -120,11 +229,17 @@ export default function PeoplePage() {
 
     // set msg
     if (response.ok) {
-      //setPeople((prev) => prev.filter((person) => person.id !== personId));
-      await fetchPatients();
-      setMsg("Patient deleted successfully");
+      await fetchPeople();
+      roleTab === "patient"
+        ? setMsg("Patient deleted successfully")
+        : setMsg("Staff deleted successfully");
+      setShowDeleteConfirm(false);
     } else {
-      setMsg("Failed to delete patient");
+      roleTab === "patient"
+        ? setMsg("Failed to delete patient")
+        : setMsg("Failed to delete staff");
+      setShowDeleteConfirm(false);
+      setSelectedPerson(null);
     }
     setShowMsg(true);
 
@@ -143,7 +258,7 @@ export default function PeoplePage() {
     setStatus(0);
     setNotes("");
 
-    setFormTitle("Add New Patient");
+    setFormTitle("Add New Person");
     setShowAddForm(true);
   };
 
@@ -157,17 +272,53 @@ export default function PeoplePage() {
     setStatus(person.status === "Active" ? 0 : 1);
     setNotes(person.notes);
 
-    setFormTitle("Edit Existing Patient");
+    setFormTitle("Edit Existing Person");
     setShowAddForm(true);
   };
 
   return (
     <div className="people">
       <div className="people-header">
-        <div>View and manage all people in the system.</div>
+        <div className="people-header-element">
+          View and manage all people in the system.
+          <div className="people-role-search">
+            {/* role selection - patient in default */}
+            <div className="people-role-tabs">
+              <button
+                type="button"
+                className={
+                  roleTab === "patient" ? "role-tab active" : "role-tab"
+                }
+                onClick={() => setRoleTab("patient")}
+              >
+                Patient
+              </button>
+              <button
+                type="button"
+                className={roleTab === "staff" ? "role-tab active" : "role-tab"}
+                onClick={() => setRoleTab("staff")}
+              >
+                Staff
+              </button>
+            </div>
+
+            <Searchbar
+              placeholder="Name, email or phone"
+              value={search}
+              onChange={(e) => {
+                setSearch(e.target.value);
+                setCurrentPage(1);
+              }}
+            />
+          </div>
+        </div>
 
         <div className="people-header-buttons">
-          <button className="people-header-button" type="button">
+          <button
+            className="people-header-button"
+            type="button"
+            onClick={handleImportClick}
+          >
             <img
               className="people-header-button-img"
               src="/icons/import.svg"
@@ -175,6 +326,13 @@ export default function PeoplePage() {
             ></img>
             <div className="people-header-button-text">Import</div>
           </button>
+          <input
+            ref={fileInputRef}
+            type="file"
+            accept=".csv"
+            style={{ display: "none" }}
+            onChange={handleImportFileChange}
+          />
 
           <button
             className="people-header-button"
@@ -257,6 +415,33 @@ export default function PeoplePage() {
                     </label>
                   </div>
                 )}
+                {!isEditMode && (
+                  <div className="app-form-input-row">
+                    <div className="app-form-input-row-label">Role:</div>
+                    <label>
+                      <input
+                        type="radio"
+                        name="role"
+                        value={"patient"}
+                        checked={role === "patient"}
+                        onChange={() => setRole("patient")}
+                      />
+                      Patient
+                    </label>
+
+                    <label>
+                      <input
+                        type="radio"
+                        name="role"
+                        value={"staff"}
+                        checked={role === "staff"}
+                        onChange={() => setRole("staff")}
+                      />
+                      Staff
+                    </label>
+                  </div>
+                )}
+
                 <div className="app-form-input-row">
                   <div className="app-form-input-row-label">Notes:</div>
                   <input
@@ -277,18 +462,47 @@ export default function PeoplePage() {
             </div>
           )}
           {showMsg && <div className="toast-message">{msg}</div>}
+          {showDeleteConfirm && (
+            <DeleteConfirm
+              text={`Are you sure you want to delete ${selectedPerson.name}? This will also delete all related sessions.`}
+              onCancel={handleCancelDelete}
+              onConfirm={handleDelete}
+            ></DeleteConfirm>
+          )}
         </div>
       </div>
 
-      <PeopleTable
-        className="people-table"
-        people={people}
-        onDelete={handleDelete}
-        onEdit={openEditForm}
-        currentPage={currentPage}
-        totalPages={totalPages}
-        setCurrentPage={setCurrentPage}
-      ></PeopleTable>
+      <PeopleFilterBar
+        onFilterStatus={(value) => {
+          setFilterStatus(value);
+          setCurrentPage(1);
+        }}
+        filterStatus={filterStatus}
+        onFilterName={(value) => {
+          setFilterName(value);
+          setCurrentPage(1);
+        }}
+        filterName={filterName}
+        onClear={handleFilterClear}
+      ></PeopleFilterBar>
+
+      {loading ? (
+        <div className="state-message">Loading people...</div>
+      ) : error ? (
+        <div className="state-message state-error">{error}</div>
+      ) : people.length === 0 ? (
+        <div className="state-message">No people found.</div>
+      ) : (
+        <PeopleTable
+          className="people-table"
+          people={people}
+          onDelete={handleOpenDeleteConfirm}
+          onEdit={openEditForm}
+          currentPage={currentPage}
+          totalPages={totalPages}
+          setCurrentPage={setCurrentPage}
+        ></PeopleTable>
+      )}
     </div>
   );
 }
